device/peripheral/mpu: Gyroscope minimal integration implementation

+++ device/peripheral/mpu.h
@@ -14,7 +14,7 @@ extern "C" {
 uint32_t mpu_init(void *i2c);
 void mpu_usb_hid(usb_hid_if_t *hid, usb_hid_if_t *hid_mouse);
 volatile int16_t *mpu_accel();
-volatile int16_t *mpu_gyro();
+volatile int32_t *mpu_gyro();
 volatile int16_t *mpu_accel_avg();
 volatile int16_t *mpu_gyro_avg();
 uint32_t mpu_cnt();
